requirements.txt
displayhatmini
Pillow
python-dotenv
requests
